refactored constants and utils
