+++ apps/studio/components/interfaces/TableGridEditor/TableConnections/TableConnectionsDiagram.constants.ts
